requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
backcall==0.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.3
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
geatpy==2.7.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icoextract==0.2.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.5.0
llvmlite==0.47.0
lxml==6.1.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
networkx==3.4.2
numba==0.65.0
numpy==1.26.4
olefile==0.47
python-multipart==0.0.20
optype==0.9.3
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
passlib==1.7.4
pefile==2024.8.26
pickleshare==0.7.5
pillow==12.2.0
pipreqs==0.5.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pyaml==26.2.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.7.2
scikit-optimize==0.10.2
scipy==1.15.3
scipy-stubs==1.15.3.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
yarg==0.1.9
